requirements.txt
torch==2.0.1
torchvision==0.15.2
open-clip-torch==2.20.0
huggingface-hub==0.16.4
tokenizers==0.13.3
opencv-python-headless==4.8.0.76
numpy==1.24.4
pillow==10.0.0
scipy==1.12
timm==0.9.5
matplotplib==3.8
wandb==0.17.0
pandas==2.2.2
scikit-learn==1.5.0
matplotlib==3.8.0
plotly==5.22.0
imageio==2.34.1
scikit-image==0.23.2
faiss-gpu==1.7.2
pytorch-metric-learning==2.5.0
torchsummary==1.5.1
transformers==4.41.1
einops==0.8.0